+++ src/main/java/org/leisureup/travel/internal/travel/dto/request/CreateTravelRequest.java
@@ -1,4 +1,4 @@
-package org.leisureup.travel.internal.travel.dto;
+package org.leisureup.travel.internal.travel.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
@@ -6,7 +6,7 @@ import org.leisureup.travel.internal.travel.domain.Travel;
 
 @AllArgsConstructor
 @NoArgsConstructor
-public class CreateTravelDto {
+public class CreateTravelRequest {
     private String travelName;
 
     public Travel toEntity() {
